+++ concurrency.go
@@ -2,26 +2,24 @@ package firehose
 
 import (
 	"sync"
-
-	pbeth "github.com/streamingfast/firehose-ethereum/types/pb/sf/ethereum/type/v2"
 )
 
 type ConcurrentFlushQueue struct {
 	bufferSize int
 
 	startSignal    chan uint64
-	jobQueue       chan *pbeth.Block
-	printBlockFunc func(block *pbeth.Block)
+	jobQueue       chan *blockOutput
+	printBlockFunc func(*blockOutput)
 	flushFunc      func()
 
 	jobWG     sync.WaitGroup
 	closeOnce sync.Once
 }
 
-func NewConcurrentFlushQueue(bufferSize int, printBlockFunc func(*pbeth.Block), flushFunc func()) *ConcurrentFlushQueue {
+func NewConcurrentFlushQueue(bufferSize int, printBlockFunc func(*blockOutput), flushFunc func()) *ConcurrentFlushQueue {
 	return &ConcurrentFlushQueue{
 		startSignal:    make(chan uint64, 1),
-		jobQueue:       make(chan *pbeth.Block, bufferSize),
+		jobQueue:       make(chan *blockOutput, bufferSize),
 		bufferSize:     bufferSize,
 		printBlockFunc: printBlockFunc,
 		flushFunc:      flushFunc,
@@ -35,8 +33,8 @@ func (q *ConcurrentFlushQueue) Start() {
 	go q.worker()
 }
 
-func (q *ConcurrentFlushQueue) Push(block *pbeth.Block) {
-	q.jobQueue <- block
+func (q *ConcurrentFlushQueue) Push(out *blockOutput) {
+	q.jobQueue <- out
 }
 
 // Close signals the worker to shut down and waits for it to finish
@@ -52,8 +50,8 @@ func (q *ConcurrentFlushQueue) Close() {
 // worker listens for blocks and serializes them sequentially
 func (q *ConcurrentFlushQueue) worker() {
 	defer q.jobWG.Done()
-	for block := range q.jobQueue {
-		q.printBlockFunc(block)
+	for out := range q.jobQueue {
+		q.printBlockFunc(out)
 		q.flushFunc()
 	}
 }